src/requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.2
colorama==0.4.3
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
future==0.18.2
glcontext==2.2.0
importlib-metadata==1.6.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
llvmlite==0.32.1
lmdb==0.98
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
moderngl==5.6.1
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.6
nibabel==3.1.0
notebook==6.0.3
numba==0.49.1
numpy==1.18.1
olefile==0.46
opencv-python==4.2.0.34
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.1.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psutil==5.7.0
pyarrow==0.17.1
Pygments==2.6.1
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrr==0.10.3
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.7.4
QtPy==1.9.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
sip==4.19.13
six==1.15.0
statsmodels==0.11.1
terminado==0.8.3
testpath==0.4.4
torch==1.5.0
torchtoolbox==0.1.4.1
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.1.0
zope.interface==5.1.0
